fix retours de Quentin
