+++ confiture-web-app/src/components/tiptap/extensions/MarkdownExtension.ts
@@ -1,0 +1,8 @@
+import { Markdown } from "@tiptap/markdown";
+
+export const MarkdownExtension =
+  Markdown.configure({
+    markedOptions: {
+      async: false
+    }
+  });
