+++ confiture-web-app/src/components/tiptap/extensions/StarterKitExtensions.ts
@@ -1,0 +1,5 @@
+import StarterKit from "@tiptap/starter-kit";
+
+export function getStarterKitExtensions(options: any) {
+  return StarterKit.configure(options);
+}
